fix(cmd): move Temporal namespace creation to cmd/init
